pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audio-diagnosis"
version = "0.2.0"
description = "Production-ready CRNN lung sound diagnosis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="Auto"}]
dependencies = [
  "librosa>=0.10.1",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "scikit-learn>=1.3.0",
  "torch>=2.1.0",
  "torchaudio>=2.1.0",
  "tqdm>=4.66.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.13.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.23.0",
  "pydantic>=2.0.0",
  "python-multipart>=0.0.9",
  "openpyxl>=3.1.0",
  "pyyaml>=6.0",
  "soundfile>=0.12.1",
  "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = ["ipykernel", "pytest", "black", "ruff", "mypy", "httpx", "pytest-cov"]

[tool.setuptools.packages.find]
where=["src"]
